14.09.2023. Work with subscribe serializer.

+++ backend/foods/views.py
@@ -1,5 +1,5 @@
 from django.shortcuts import get_object_or_404
-from rest_framework import mixins, status, viewsets
+from rest_framework import mixins, serializers, status, viewsets
 from rest_framework.decorators import action
 from rest_framework.response import Response
 
@@ -32,7 +32,12 @@ class RecipeViewSet(viewsets.ModelViewSet):
             }
         serializer = FavoriteSerializer(data=data)
         if self.request.method == 'DELETE':
-            serializer.is_valid_on_delete(data)
+            if not Favorites.objects.filter(
+                user=user, recipes=recipes
+            ).exists():
+                raise serializers.ValidationError(
+                    {'error': 'Вы не подписаны на этот рецепт.'}
+                )
             Favorites.objects.filter(user=user, recipes=recipes).delete()
             return Response(status=status.HTTP_204_NO_CONTENT)
         serializer.is_valid(raise_exception=True)
